requirements.txt
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
black==20.8b1
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coverage==5.3
cryptography==3.2.1
distlib==0.3.1
docutils==0.16
-e git+https://github.com/tausiftt5238/cs5204_election_guard.git@89e6a7588ea2ff223987f535091e650cb08f61ee#egg=electionguard
filelock==3.0.12
future==0.18.2
gmpy2==2.0.8
hypothesis==5.41.1
idna==2.10
iniconfig==1.1.1
isort==5.6.4
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.17.0
jsons==1.3.0
jsonschema==3.2.0
keyring==21.4.0
lazy-object-proxy==1.4.3
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
mypy==0.790
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.4
packaging==20.4
pathspec==0.8.0
Pillow==8.0.1
pipenv==2020.11.4
pkginfo==1.6.1
pluggy==0.13.1
psutil==5.7.3
py==1.9.0
pycparser==2.20
pydeps==1.9.11
pydocstyle==5.1.1
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
PyYAML==5.3.1
readme-renderer==28.0
regex==2020.10.28
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
SecretStorage==3.1.2
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
stdlib-list==0.7.0
toml==0.10.2
tornado==6.1
tqdm==4.51.0
twine==3.2.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
typish==1.9.1
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
webencodings==0.5.1
wrapt==1.12.1
